requirements.txt
# Core ML and AI libraries
torch>=2.3.0
transformers>=4.42.0
openai>=1.34.0
numpy>=1.26.0
pandas>=2.2.2
scikit-learn>=1.5.0
xgboost>=2.0.0
tokenizers>=0.15.2
sentencepiece>=0.1.99
accelerate>=0.33.0
huggingface-hub>=0.24.0
peft>=0.11.0
tqdm>=4.66.3
scipy>=1.13.0

# Data processing and utilities
requests>=2.32.0
beautifulsoup4>=4.12.3
lxml>=4.9.0
python-dotenv>=1.0.0
datasets>=2.20.0
evaluate>=0.4.1

# Input processing and OCR
easyocr>=1.7.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Speech recognition
openai-whisper>=20231117
moviepy>=1.0.3

# Translation
googletrans==4.0.0rc1
xxhash>=3.4.1
google-cloud-firestore>=2.16.0

# Phase 5: Manipulation Detection & User Education
dataclasses>=0.6; python_version<"3.7"
enum34>=1.1.10; python_version<"3.4"

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter for experiments
jupyter>=1.0.0
ipykernel>=6.25.0

# Vector search & embeddings
faiss-cpu>=1.11.0
sentence-transformers>=5.1.0

# Multimedia processing
opencv-python>=4.8.0
easyocr>=1.7.0
openai-whisper>=20231117

# Web scraping & text processing
trafilatura>=7.0.0
tldextract>=4.0.0
dateparser>=1.1.0
langdetect>=1.0.9
